+++ classes/shaman.lua
@@ -1,5 +1,4 @@
 local AddonName, SAO = ...
-local Module = "shaman"
 
 -- Optimize frequent calls
 local CombatLogGetCurrentEventInfo = CombatLogGetCurrentEventInfo
@@ -129,7 +128,7 @@ local function rollingThunderCombatCheck(combat)
         if RollingThunderHandler.glowtimer then
             RollingThunderHandler.glowtimer:Cancel();
             RollingThunderHandler.glowtimer = nil;
-            end
+        end
         checkRollingThunderRuneAndLightningSieldStacks();
     end
 end
@@ -162,12 +161,12 @@ local function registerClass(self)
             talent = 16164, -- Elemental Focus (talent)
             overlays = {
                 [SAO.TBC+SAO.WRATH] = {
-                    { stacks = 1, texture = "echo_of_the_elements", position = "Left", scale = 1, option = false },
-                    { stacks = 2, texture = "echo_of_the_elements", position = "Left + Right (Flipped)", scale = 1, option = { setupStacks = 0, testStacks = 2 } },
+                    { stacks = 1, texture = "echo_of_the_elements", position = "Left", scale = 1, pulse = false, option = false },
+                    { stacks = 2, texture = "echo_of_the_elements", position = "Left + Right (Flipped)", scale = 1, pulse = false, option = { setupStacks = 0, testStacks = 2 } },
                 },
                 [SAO.CATA] = {
-                    { stacks = 1, texture = "genericarc_05", position = "Left", scale = 1.5, option = false },
-                    { stacks = 2, texture = "genericarc_05", position = "Left + Right (Flipped)", scale = 1.5, option = { setupStacks = 0, testStacks = 2 } },
+                    { stacks = 1, texture = "genericarc_05", position = "Left", scale = 1.5, pulse = false, option = false },
+                    { stacks = 2, texture = "genericarc_05", position = "Left + Right (Flipped)", scale = 1.5, pulse = false, option = { setupStacks = 0, testStacks = 2 } },
                 }
             },
         }
@@ -187,6 +186,65 @@ local function registerClass(self)
         }
     );
 
+    -- Tidal Waves
+    local greaterHealingWave = 77472;
+    local healingSurge = 8004;
+    local healingWave = 331;
+    local lesserHealingWave = 8004; -- Renamed Healing Surge in Cataclysm; keep the former name to make the effect easier to design
+    local tidalWavesBuff = self.IsSoD() and 432041 or 53390;
+    local tidalWavesTalent = self.IsSoD() and 432233 or 51562;
+    self:CreateEffect(
+        "tidal_waves",
+        SAO.SOD + SAO.WRATH + SAO.CATA,
+        tidalWavesBuff,
+        "aura",
+        {
+            talent = tidalWavesTalent,
+            overlays = {
+                { stacks = 1, texture = "high_tide", position = "Left (CCW)", scale = 0.8, option = false },
+                { stacks = 2, texture = "high_tide", position = "Left (CCW)", scale = 0.8, option = false },
+                { stacks = 2, texture = "high_tide", position = "Right (CW)", scale = 0.8, option = { setupStacks = 0, testStacks = 2 } },
+            },
+            buttons = {
+                [SAO.SOD+SAO.WRATH] = { lesserHealingWave, healingWave },
+                [SAO.CATA] = { greaterHealingWave, healingWave, healingSurge },
+            },
+        }
+    );
+
+    -- Maelstrom Weapon
+    local lightningBolt = 403;
+    local chainLightning = 421;
+    local chainHeal = 1064;
+    local healingRain = 73920;
+    local hex = 51514;
+    local lavaBurstSoD = 408490;
+    local maelstromWeaponBuff = self.IsSoD() and 408505 or 53817;
+    local maelstromWeaponTalent = self.IsSoD() and 408498 or 51528;
+    local maelstromWeaponScale = self.IsSoD() and 0.8 or 1;
+    self:CreateEffect(
+        "maelstrom_weapon",
+        SAO.SOD + SAO.WRATH + SAO.CATA,
+        maelstromWeaponBuff,
+        "aura",
+        {
+            talent = maelstromWeaponTalent,
+            overlays = {
+                { stacks = 1, texture = "maelstrom_weapon_1", position = "Top", scale = maelstromWeaponScale, pulse = false, option = false },
+                { stacks = 2, texture = "maelstrom_weapon_2", position = "Top", scale = maelstromWeaponScale, pulse = false, option = false },
+                { stacks = 3, texture = "maelstrom_weapon_3", position = "Top", scale = maelstromWeaponScale, pulse = false, option = false },
+                { stacks = 4, texture = "maelstrom_weapon_4", position = "Top", scale = maelstromWeaponScale, pulse = false, option = { setupStacks = 0, testStacks = 4, subText = self:NbStacks(1,4) } },
+                { stacks = 5, texture = "maelstrom_weapon"  , position = "Top", scale = maelstromWeaponScale, pulse = true , option = true },
+            },
+            buttons = {
+                default = { stacks = 5 },
+                [SAO.SOD] =   { lightningBolt, chainLightning, lesserHealingWave,                     healingWave, chainHeal,                   lavaBurstSoD },
+                [SAO.WRATH] = { lightningBolt, chainLightning, lesserHealingWave,                     healingWave, chainHeal,              hex },
+                [SAO.CATA] =  { lightningBolt, chainLightning, healingSurge,      greaterHealingWave, healingWave, chainHeal, healingRain, hex },
+            },
+        }
+    );
+
     if self.IsCata() then
         -- Initializing Rolling Thunder handler for Fulmination in Cataclysm
         if (not RollingThunderHandler.initialized) then
@@ -198,81 +256,16 @@ local function registerClass(self)
             local pulse = lightningShieldStacks == 9;
             self:RegisterAura(auraName, lightningShieldStacks, RollingThunderHandler.fakeSpellID, "fulmination", "Top", scale, 255, 255, 255, pulse, RollingThunderHandler.earthShockSpells);
         end
-        -- Maelstrom Weapon
-        local lightningBolt = 403;
-        local chainLightning = 421;
-        local healingSurge = 8004;
-        local healingWave = 331;
-        local chainHeal = 1064;
-        local greaterHealingWave = 77472;
-        local healingRain = 73920;
-        local hex = 51514;
-        local maelstromSpells = {
-            (GetSpellInfo(lightningBolt)),
-            (GetSpellInfo(chainLightning)),
-            (GetSpellInfo(healingSurge)),
-            (GetSpellInfo(greaterHealingWave)),
-            (GetSpellInfo(healingWave)),
-            (GetSpellInfo(chainHeal)),
-            (GetSpellInfo(healingRain)),
-            (GetSpellInfo(hex)),
-        }
-        self:RegisterAura("maelstrom_weapon_1", 1, 53817, "maelstrom_weapon_1", "Top", 1, 255, 255, 255, false);
-        self:RegisterAura("maelstrom_weapon_2", 2, 53817, "maelstrom_weapon_2", "Top", 1, 255, 255, 255, false);
-        self:RegisterAura("maelstrom_weapon_3", 3, 53817, "maelstrom_weapon_3", "Top", 1, 255, 255, 255, false);
-        self:RegisterAura("maelstrom_weapon_4", 4, 53817, "maelstrom_weapon_4", "Top", 1, 255, 255, 255, false);
-        self:RegisterAura("maelstrom_weapon_5", 5, 53817, "maelstrom_weapon", "Top", 1, 255, 255, 255, true, maelstromSpells);
-
-        -- Tidal Waves
-        local tidalSpells = {
-            (GetSpellInfo(greaterHealingWave)),
-            (GetSpellInfo(healingWave)),
-            (GetSpellInfo(healingSurge)),
-        }
-        self:RegisterAura("tidal_waves_1", 1, 53390, "high_tide", "Left (CCW)", 0.8, 255, 255, 255, true, tidalSpells);
-        self:RegisterAura("tidal_waves_2", 2, 53390, "high_tide", "Left (CCW)", 0.8, 255, 255, 255, true, tidalSpells);
-        self:RegisterAura("tidal_waves_2", 2, 53390, "high_tide", "Right (CW)", 0.8, 255, 255, 255, true); -- no need to re-glow tidalSpells for right texture
     end
 
-
     if self.IsWrath() then
-        -- Maelstrom Weapon
-        local lightningBolt = 403;
-        local chainLightning = 421;
-        local lesserHealingWave = 8004;
-        local healingWave = 331;
-        local chainHeal = 1064;
-        local hex = 51514;
-        local maelstromSpells = {
-            (GetSpellInfo(lightningBolt)),
-            (GetSpellInfo(chainLightning)),
-            (GetSpellInfo(lesserHealingWave)),
-            (GetSpellInfo(healingWave)),
-            (GetSpellInfo(chainHeal)),
-            (GetSpellInfo(hex)),
-        }
-        self:RegisterAura("maelstrom_weapon_1", 1, 53817, "maelstrom_weapon_1", "Top", 1, 255, 255, 255, false);
-        self:RegisterAura("maelstrom_weapon_2", 2, 53817, "maelstrom_weapon_2", "Top", 1, 255, 255, 255, false);
-        self:RegisterAura("maelstrom_weapon_3", 3, 53817, "maelstrom_weapon_3", "Top", 1, 255, 255, 255, false);
-        self:RegisterAura("maelstrom_weapon_4", 4, 53817, "maelstrom_weapon_4", "Top", 1, 255, 255, 255, false);
-        self:RegisterAura("maelstrom_weapon_5", 5, 53817, "maelstrom_weapon", "Top", 1, 255, 255, 255, true, maelstromSpells);
-
-        -- Tidal Waves
-        local tidalSpells = {
-            (GetSpellInfo(lesserHealingWave)),
-            (GetSpellInfo(healingWave)),
-        }
-        self:RegisterAura("tidal_waves_1", 1, 53390, "high_tide", "Left (CCW)", 0.8, 255, 255, 255, true, tidalSpells);
-        self:RegisterAura("tidal_waves_2", 2, 53390, "high_tide", "Left (CCW)", 0.8, 255, 255, 255, true, tidalSpells);
-        self:RegisterAura("tidal_waves_2", 2, 53390, "high_tide", "Right (CW)", 0.8, 255, 255, 255, true); -- no need to re-glow tidalSpells for right texture
-
         -- Healing Trance / Soul Preserver
         self:RegisterAuraSoulPreserver("soul_preserver_shaman", 60515); -- 60515 = Shaman buff
     end
 
     if self.IsEra() and not self.IsSoD() then
         -- On non-SoD Era, Elemental Focus is simply displayed Left and Right
-        self:RegisterAura("elemental_focus", 0, 16246, "echo_of_the_elements", "Left + Right (Flipped)", 1, 255, 255, 255, true);
+        self:RegisterAura("elemental_focus", 0, 16246, "echo_of_the_elements", "Left + Right (Flipped)", 1, 255, 255, 255, false);
     end
 
     if self.IsSoD() then
@@ -286,33 +279,13 @@ local function registerClass(self)
         self:RegisterAura("molten_blast", 0, moltenBlastSoD, "impact", "Top", 0.8, 255, 255, 255, true, { moltenBlastSoD }, true);
         self:RegisterCounter("molten_blast");
 
-        -- Maelstrom Weapon & Power Surge
-        local maelstromSoDBuff = 408505;
-        local lightningBolt = 403;
-        local chainLightning = 421;
-        local lesserHealingWave = 8004;
-        local healingWave = 331;
-        local chainHeal = 1064;
-        local lavaBurstSoD = 408490;
+        -- Power Surge
         local powerSurgeSoDBuff = 415105;
-        local maelstromSpells = {
-            (GetSpellInfo(lightningBolt)),
-            (GetSpellInfo(chainLightning)),
-            (GetSpellInfo(lesserHealingWave)),
-            (GetSpellInfo(healingWave)),
-            (GetSpellInfo(chainHeal)),
-            (GetSpellInfo(lavaBurstSoD)),
-        }
         local powerSurgeSpells = {
             (GetSpellInfo(chainLightning)),
             (GetSpellInfo(chainHeal)),
             (GetSpellInfo(lavaBurstSoD)),
         }
-        self:RegisterAura("maelstrom_weapon_sod_1", 1, maelstromSoDBuff, "maelstrom_weapon_1", "Top", 0.8, 255, 255, 255, false);
-        self:RegisterAura("maelstrom_weapon_sod_2", 2, maelstromSoDBuff, "maelstrom_weapon_2", "Top", 0.8, 255, 255, 255, false);
-        self:RegisterAura("maelstrom_weapon_sod_3", 3, maelstromSoDBuff, "maelstrom_weapon_3", "Top", 0.8, 255, 255, 255, false);
-        self:RegisterAura("maelstrom_weapon_sod_4", 4, maelstromSoDBuff, "maelstrom_weapon_4", "Top", 0.8, 255, 255, 255, false);
-        self:RegisterAura("maelstrom_weapon_sod_5", 5, maelstromSoDBuff, "maelstrom_weapon", "Top", 0.8, 255, 255, 255, true, maelstromSpells);
 
         -- If Power Surge is enabled but not Elemental Focus, display Power Surge Left and Right
         -- If Elemental Focus is enabled but not Power Surge, display Elemental Focus Left and Right
@@ -343,63 +316,36 @@ local function registerClass(self)
 
         self:RegisterAura("power_surge_sod", 0, powerSurgeSoDBuff, "imp_empowerment", "Left", 1, 255, 255, 255, true, powerSurgeSpells);
         self:RegisterAura("power_surge_sod", 0, powerSurgeSoDBuff, powerSurgeRightTextureFunc, "Right (Flipped)", 1, 255, 255, 255, true, powerSurgeSpells);
-        self:RegisterAura("elemental_focus", 0, elementalFocusBuff, elementalFocusLeftTextureFunc, "Left", 1, 255, 255, 255, true);
-        self:RegisterAura("elemental_focus", 0, elementalFocusBuff, "echo_of_the_elements", "Right (Flipped)", 1, 255, 255, 255, true);
+        self:RegisterAura("elemental_focus", 0, elementalFocusBuff, elementalFocusLeftTextureFunc, "Left", 1, 255, 255, 255, false);
+        self:RegisterAura("elemental_focus", 0, elementalFocusBuff, "echo_of_the_elements", "Right (Flipped)", 1, 255, 255, 255, false);
         for lightningShieldStacks=7,9 do
             local auraName = "rolling_thunder_"..lightningShieldStacks;
             local scale = 0.5 + 0.1 * (lightningShieldStacks - 6); -- 60%, 70%, 80% for Season of Discovery
             local pulse = lightningShieldStacks == 9;
             self:RegisterAura(auraName, lightningShieldStacks, RollingThunderHandler.fakeSpellID, "fulmination", "Top", scale, 255, 255, 255, pulse, RollingThunderHandler.earthShockSpells);
         end
-        -- Tidal Waves SoD
-        local tidalSpells = {
-            GetSpellInfo(lesserHealingWave),
-            GetSpellInfo(healingWave),
-        }
-        self:RegisterAura("tidal_waves_1_sod", 1, 432041, "high_tide", "Left (CCW)", 0.8, 255, 255, 255, true, tidalSpells);
-        self:RegisterAura("tidal_waves_2_sod", 2, 432041, "high_tide", "Left (CCW)", 0.8, 255, 255, 255, true, tidalSpells);
-        self:RegisterAura("tidal_waves_2_sod", 2, 432041, "high_tide", "Right (CW)", 0.8, 255, 255, 255, true); -- no need to re-glow tidalSpells for right texture
     end
 end
 
 local function loadOptions(self)
-    local lightningBolt = 403;
     local chainLightning = 421;
-    local lesserHealingWave = 8004;
-    local healingWave = 331;
     local chainHeal = 1064;
-    local hex = 51514;
-
-    local maelstromWeaponBuff = 53817;
-    local maelstromWeaponTalent = 51528;
 
     local elementalFocusBuff = 16246;
     local elementalFocusTalent = 16164;
 
-    local tidalWavesBuff = 53390;
-    local tidalWavesTalent = 51562;
-
     -- Season of Discovery
     local moltenBlastSoD = 425339;
-    local maelstromSoDBuff = 408505;
-    local maelstromSoD = 408498;
     local lavaBurstSoD = 408490;
     local powerSurgeSoDBuff = 415105;
     local powerSurgeSoD = 415100;
     local lightningShield = 324;
     local rollingThunderSoD = 432056;
     local earthShock = 8042;
-    local tidalWavesSoDBuff = 432041;
-    local tidalWavesSoDTalent = 432233;
 
     --Cataclysm
     local fulminationTalentCata = 88766;
-    local greaterHealingWave = 77472;
-    local healingSurge = 8004;
-    local healingRain = 73920;
 
-    local oneToFourStacks = self:NbStacks(1, 4);
-    local fiveStacks = self:NbStacks(5);
     local sevenToNineStacks = self:NbStacks(7, 9);
     local sixToNineStacks = self:NbStacks(6, 9);
 
@@ -413,63 +359,25 @@ local function loadOptions(self)
         self:AddOverlayOption(fulminationTalentCata, lightningShield, 7, nil, nil, nil, RollingThunderHandler.fakeSpellID);
         self:AddOverlayOption(fulminationTalentCata, lightningShield, 8, nil, nil, nil, RollingThunderHandler.fakeSpellID);
         self:AddOverlayOption(fulminationTalentCata, lightningShield, 9, nil, nil, nil, RollingThunderHandler.fakeSpellID);
-        self:AddOverlayOption(maelstromWeaponTalent, maelstromWeaponBuff, 0, oneToFourStacks, nil, 4); -- setup any stacks, test with 4 stacks
-        self:AddOverlayOption(maelstromWeaponTalent, maelstromWeaponBuff, 5); -- setup 5 stacks
-        self:AddOverlayOption(tidalWavesTalent, tidalWavesBuff, 0, nil, nil, 2); -- setup any stacks, test with 2 stacks
-    end 
+    end
     if self.IsWrath() then
-        self:AddOverlayOption(maelstromWeaponTalent, maelstromWeaponBuff, 0, oneToFourStacks, nil, 4); -- setup any stacks, test with 4 stacks
-        self:AddOverlayOption(maelstromWeaponTalent, maelstromWeaponBuff, 5); -- setup 5 stacks
-        self:AddOverlayOption(tidalWavesTalent, tidalWavesBuff, 0, nil, nil, 2); -- setup any stacks, test with 2 stacks
         self:AddSoulPreserverOverlayOption(60515); -- 60515 = Shaman buff
     elseif self.IsSoD() then
         self:AddOverlayOption(powerSurgeSoD, powerSurgeSoDBuff);
         self:AddOverlayOption(moltenBlastSoD, moltenBlastSoD);
-        self:AddOverlayOption(maelstromSoD, maelstromSoDBuff, 0, oneToFourStacks, nil, 4); -- setup any stacks, test with 4 stacks
-        self:AddOverlayOption(maelstromSoD, maelstromSoDBuff, 5); -- setup 5 stacks
         self:AddOverlayOption(rollingThunderSoD, lightningShield, 7, nil, nil, nil, RollingThunderHandler.fakeSpellID);
         self:AddOverlayOption(rollingThunderSoD, lightningShield, 8, nil, nil, nil, RollingThunderHandler.fakeSpellID);
         self:AddOverlayOption(rollingThunderSoD, lightningShield, 9, nil, nil, nil, RollingThunderHandler.fakeSpellID);
-        self:AddOverlayOption(tidalWavesSoDTalent, tidalWavesSoDBuff, 0, nil, nil, 2); -- setup any stacks, test with 2 stacks
     end
 
     if self.IsCata() then
         self:AddGlowingOption(fulminationTalentCata, lightningShield, earthShock, sixToNineStacks);
-        self:AddGlowingOption(maelstromWeaponTalent, maelstromWeaponBuff, lightningBolt, fiveStacks);
-        self:AddGlowingOption(maelstromWeaponTalent, maelstromWeaponBuff, chainLightning, fiveStacks);
-        self:AddGlowingOption(maelstromWeaponTalent, maelstromWeaponBuff, healingSurge, fiveStacks);
-        self:AddGlowingOption(maelstromWeaponTalent, maelstromWeaponBuff, greaterHealingWave, fiveStacks);
-        self:AddGlowingOption(maelstromWeaponTalent, maelstromWeaponBuff, healingWave, fiveStacks);
-        self:AddGlowingOption(maelstromWeaponTalent, maelstromWeaponBuff, chainHeal, fiveStacks);
-        self:AddGlowingOption(maelstromWeaponTalent, maelstromWeaponBuff, healingRain, fiveStacks);
-        self:AddGlowingOption(maelstromWeaponTalent, maelstromWeaponBuff, hex, fiveStacks);
-        self:AddGlowingOption(tidalWavesTalent, tidalWavesBuff, greaterHealingWave);
-        self:AddGlowingOption(tidalWavesTalent, tidalWavesBuff, healingWave);
-        self:AddGlowingOption(tidalWavesTalent, tidalWavesBuff, healingSurge);
-    end 
-    if self.IsWrath() then
-        self:AddGlowingOption(maelstromWeaponTalent, maelstromWeaponBuff, lightningBolt, fiveStacks);
-        self:AddGlowingOption(maelstromWeaponTalent, maelstromWeaponBuff, chainLightning, fiveStacks);
-        self:AddGlowingOption(maelstromWeaponTalent, maelstromWeaponBuff, lesserHealingWave, fiveStacks);
-        self:AddGlowingOption(maelstromWeaponTalent, maelstromWeaponBuff, healingWave, fiveStacks);
-        self:AddGlowingOption(maelstromWeaponTalent, maelstromWeaponBuff, chainHeal, fiveStacks);
-        self:AddGlowingOption(maelstromWeaponTalent, maelstromWeaponBuff, hex, fiveStacks);
-        self:AddGlowingOption(tidalWavesTalent, tidalWavesBuff, lesserHealingWave);
-        self:AddGlowingOption(tidalWavesTalent, tidalWavesBuff, healingWave);
     elseif self.IsSoD() then
         self:AddGlowingOption(nil, moltenBlastSoD, moltenBlastSoD);
         self:AddGlowingOption(powerSurgeSoD, powerSurgeSoDBuff, chainLightning);
         self:AddGlowingOption(powerSurgeSoD, powerSurgeSoDBuff, chainHeal);
         self:AddGlowingOption(powerSurgeSoD, powerSurgeSoDBuff, lavaBurstSoD);
-        self:AddGlowingOption(maelstromSoD, maelstromSoDBuff, lightningBolt, fiveStacks);
-        self:AddGlowingOption(maelstromSoD, maelstromSoDBuff, chainLightning, fiveStacks);
-        self:AddGlowingOption(maelstromSoD, maelstromSoDBuff, lesserHealingWave, fiveStacks);
-        self:AddGlowingOption(maelstromSoD, maelstromSoDBuff, healingWave, fiveStacks);
-        self:AddGlowingOption(maelstromSoD, maelstromSoDBuff, chainHeal, fiveStacks);
-        self:AddGlowingOption(maelstromSoD, maelstromSoDBuff, lavaBurstSoD, fiveStacks);
         self:AddGlowingOption(rollingThunderSoD, lightningShield, earthShock, sevenToNineStacks);
-        self:AddGlowingOption(tidalWavesSoDTalent, tidalWavesSoDBuff, lesserHealingWave);
-        self:AddGlowingOption(tidalWavesSoDTalent, tidalWavesSoDBuff, healingWave);
     end
 end